ci: run both suites on every push, and clean the lint they will enforce
There was no CI. 380 tests that nobody runs on push are documentation, not a
safety net, and every defect fixed this week was found by hand.

The workflow is two jobs on purpose: the engine suite and the front-end suite
fail for unrelated reasons, so a red badge should already say which one broke.
Neither needs a secret or a service -- the trained model is committed and the
front-end tests bring their own fake server. The python job runs exactly the
three commands the README tells a reader to run.

Latest ruff reports twelve findings the pinned-by-nothing local tool did not.
Ten are fixed. Two are declined in place, with the reason written down:

  * __all__ is grouped to mirror the two imports above it; alphabetising it
    would interleave board vocabulary with search vocabulary.
  * "alpha = max(alpha, best_score)" reads better than the compare-and-store
    it replaces, but it sits on the innermost line of negamax and costs
    18.5 ns against 6.5 ns (CPython 3.14, 2e6 iterations, best of 5).

One of the ten was a real test defect: the store-eviction test asserted
pytest.raises(Exception), which also passes when store.get raises an
AttributeError from a typo. It now names HTTPException and checks for a 404.

ruff is pinned to a minor range in the dev extras so the lint cannot start
failing because the tool moved underneath it.

+++ src/connect4/dataset.py
@@ -318,8 +318,8 @@ def stratified_split(
         rng.shuffle(indices)
 
         n = len(indices)
-        n_val = int(round(n * val_fraction))
-        n_test = int(round(n * test_fraction))
+        n_val = round(n * val_fraction)
+        n_test = round(n * test_fraction)
 
         splits["val"].append(indices[:n_val])
         splits["test"].append(indices[n_val : n_val + n_test])
